fix: variant type issues  (node-opcua#1088)
 - this commit fixes some variant type mismatch that were
   not highlighted before node-opcua@2.60.2.
 - node-opcua@2.60.2 onwards now apply a strictly type checking
   to help identifying mistmatch when setValueFromSource is used
   on the server side.
 - related to node-opcua#1088 issue

+++ src/serveraddressspace/serveraddressspace.ts
@@ -103,12 +103,12 @@ export const createOwnServerAddressspace = async (addressSpace: AddressSpace): P
         browseName: "MySeverity",
         componentOf: dev,
         description: coerceLocalizedText("Value must be between 1000 and 100") || undefined,
-        dataType: DataType.Int32,
+        dataType: DataType.Double,
         value: {
             get: () => {
                 return new Variant({
                     value: count,
-                    dataType: DataType.Int32
+                    dataType: DataType.Double
             })},
             set: (variant: Variant) => {
                 if (variant.value > 1000 || variant.value < 100) {
